+++ Archivist/src/main/java/com/zorroa/archivist/repository/RoomDaoImpl.java
@@ -23,7 +23,7 @@ public class RoomDaoImpl extends AbstractDao implements RoomDao {
 
     private static final RowMapper<Room> MAPPER = (rs, row) -> {
         Room room = new Room();
-        room.setId(rs.getLong("pk_room"));
+        room.setId(rs.getInt("pk_room"));
         room.setName(rs.getString("str_name"));
         room.setVisible(rs.getBoolean("bool_visible"));
         return room;
